+++ src/app/(sort)/selection_sort/page.tsx
@@ -1,42 +1,10 @@
-'use client'
+import SelectionSortDetail from "./detail";
+import { Metadata } from "next";
 
-import { SelectionSortFactory, SortGraph, getGraphSize } from "@/components/canvas/sort_graph";
-import { useEffect, useState } from "react";
-import { SortControl } from "@/components/control";
-import { usePathname, useRouter } from "next/navigation";
-
+export const metadata: Metadata = {
+    title: "Selection"
+}
 
 export default function SelectionSortPage() {
-    // Because the document is undefined due to server side rendering,
-    // An useState and an useEffect hook is needed to get the size
-    // of the graph
-    let [ [svgWidth, svgHeight], setSvgSize ] = useState([0, 0]);
-    let [graph, setGraph] = useState<SortGraph | undefined>();
-    let path = usePathname();
-
-    useEffect(() => {
-        const size = document.getElementById("inputSize") as HTMLInputElement;
-        let createdGraph: SortGraph;
-        let handlerSnapshot: () => void;
-        if (size !== null && svgWidth > 0) {
-            const factory = new SelectionSortFactory(parseInt(size.value), svgWidth, svgHeight);
-            createdGraph = factory.createGraph();
-            setGraph(createdGraph);
-            handlerSnapshot = SortControl.addHandler(factory, createdGraph, setGraph);
-
-        }
-        getGraphSize(setSvgSize);
-
-        return () => {
-            SortControl.removeStartHandler(handlerSnapshot);
-        }
-    }, [svgWidth, svgHeight]);
-
-    console.log(graph)
-
-    return (
-        <>
-            {graph !== undefined && graph.draw()}
-        </>
-    );
+    return <SelectionSortDetail />
 }
